# api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: pgsink
  version: 1.0.0
servers:
- url: http://localhost:8000
paths:
...
  /api/subscriptions/current/actions/add-table:
    post:
      tags:
      - Subscriptions
      summary: AddTable Subscriptions
      description: Add table to publication, relying on an import manager to schedule
        the job
      operationId: Subscriptions#AddTable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionPublishedTable'
            example:
              name: payments
              schema: public
      responses:
        "201":
          description: Created response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
              example:
                id: e32ur90j2r
                published_tables:
                - name: payments
                  schema: public
                - name: payments
                  schema: public
                - name: payments
                  schema: public
  /api/subscriptions/current/actions/stop-table:
    post:
      tags:
      - Subscriptions
      summary: StopTable Subscriptions
      description: Stop a table by removing it from the publication, and expiring
        any import jobs
      operationId: Subscriptions#StopTable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionPublishedTable'
            example:
              name: payments
              schema: public
      responses:
        "202":
          description: Accepted response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
              example:
                id: e32ur90j2r
                published_tables:
                - name: payments
                  schema: public
                - name: payments
                  schema: public
                - name: payments
                  schema: public
                - name: payments
                  schema: public
...
components:
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: string
          description: ID of subscription
          example: e32ur90j2r
        published_tables:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionPublishedTable'
          description: List of published tables
          example:
          - name: payments
            schema: public
          - name: payments
            schema: public
          - name: payments
            schema: public
          - name: payments
            schema: public
      example:
        id: e32ur90j2r
        published_tables:
        - name: payments
          schema: public
        - name: payments
          schema: public
      required:
      - id
      - published_tables
    SubscriptionPublishedTable:
      type: object
      properties:
        name:
          type: string
          description: Postgres table name
          example: payments
        schema:
          type: string
          description: Postgres table schema
          example: public
      description: Table on subscription that is published
      example:
        name: payments
        schema: public
      required:
      - schema
      - name
      - approximate_row_count
      - publication_status
      - import_status
      - import_rows_processed_total
...